requirements.txt
Django==4.2.7
django-extensions
ipython
Jinja2
django-debug-toolbar
django-admin-rangefilter
django-import-export
django-storages
django-filter~=23.5
boto3
Celery~=5.3.6
requests~=2.31.0
djangorestframework~=3.14.0
djangorestframework-xml
djangorestframework-simplejwt
drf-yasg


#flake8
flake8
flake8-builtins
flake8-print
flake8-breakpoint

Currency~=0.1.2
djangorestframework-yaml~=2.0.0
django-suit~=2.0a2

#pytest
pytest
pytest-django